fix(codegen): set placeholder env before module load so ci spec export needs no .env
